clean up demo
