--- application/extraResources/demo/day_3.socool
@@ -1,226 +1,0 @@
-import songs/std.socool
-{ f: 440, l: 19/10, g: 1, p: 0 }
-
-overtones_1 = {
-    O[
-    	(1/1, 1, 1, 4/5),
-    	(1, 1/10, 1, 0),
-      	(1/1, 0, 1, -4/5),
-  	]
-}
-
-overtones_2 = {
-    O[
-        (2, 0, 1/4, 0),
-    	(1/1, 1, 1, 3/5),
-    	(1, 1/10, 1, 0),
-      	(1/1, 0, 1, -3/5),
-        (1/2, 1/11, 1/4, 0),
-  	]
-}
-
-a_1 = {
-	Sequence [
-	    overtones_1
-        | Seq [
-            1^1, 4/5^1, 15/16^2, 
-            3/4^2, 4/5^2,
-        ],
-	] 
-}
-
-a_2 = {
-    overtones_2
-    | Seq [
-        0^1, 4/5^1, 1^1, 9/8^1 | Ta -3, 9/8^1 | Ta -5,
-        9/8^1 | Ta -9, 5/4^1, 4/3^1, 5/4^1, 0^1, 
-    ]
-}
-
-a_3 = {
-    overtones_2
-    | Seq [
-        9/8^4, 5/4^2 | Tm 15/16, 5/4^2, 4/3^1, 3/2^1,
-    ]
-}
-
-a_4 = {
-    overtones_2
-    | Seq [
-        Length 4/5 |Seq [
-            5/3^2, 5/4^1, 9/8^1, 5/4^4,
-            9/8^2, 5/6^1, 9/8^1, 1^3, 0^1,
-        ] | Repeat 2
-    ]
-}
-
-
-b_1 = {
-    Tm 1/2
-	| overtones_1
-    | Seq [
-        1^1, 15/16^1, 9/8^2, 
-        5/4^1, 9/8^1, 4/3^2,
-    ]
-    
-}
-
-b_2 = {
-    Tm 1/2
-    | overtones_2
-    | Seq [
-        0^1, 1/1^1, 5/4^1, 9/8^2, 
-        0^1, 15/16^1, 9/8^1, 15/16^1, 4/5^1, 
-    ]
-}
-
-b_3 = {
-    Tm 1/2
-    | overtones_2
-    | Seq [
-        5/6^2, 7/8^2, 15/16^3, 5/4^3 | Tm 5/8,	
-    ]
-    
-}
-
-b_4 = {
-    Tm 1/2
-    | overtones_2
-    | Seq [
-        Length 4/5 | Seq [
-            2/3^2, 3/4^2, 5/6^2, 3/4^2,
-            7/8^2, 2/3^1, 3/4^1, 4/5^2, 3/5^1, 0^1,
-        ] | Repeat 2
-    ]
-    
-}
-
-c_3 = {
-    O[
-    	(1/1, 0, 1, 4/5),
-      	(1/1, 1, 1, -4/5),
-  	]
-    | Tm 1
-    | Seq [
-        4/3^3, Tm 9/8 | Seq [5/4^1, 11/8^1, 4/3^3],
-        5/3^1, 15/8^1,
-    ]    
-}
-
-c_4 = {
-    O[
-    	(1/1, 0, 1, 1),
-      	(1/1, 1, 1, -1),
-      	(3/4, 0, 1/2, 1),
-      	(3/4, 0, 1/2, -1),
-  	]
-    | Tm 1
-    | Seq [
-        Seq [
-            2^5/2, 5/3^1/2, 3/2^1/4, 
-            5/3^9
-            | ModulateBy [
-                std.melt 
-                | Seq [
-                    Length 10, 
-                    Ta -20, Ta -40, Ta -60, Ta -80, 
-                    Ta -100
-                ]
-                | ModulateBy [Length 8, std.fade_out]
-                | ModulateBy [
-                    Length 24, 
-                    Seq [1^1, 0^1, 1^1, 0^3/2] | Repeat 6
-                ]
-            ],
-        ]
-    ]    
-}
-
-
-main = {
-    Overlay [
-        Seq [
-            a_1 | Length 5/4, 
-            a_2, 
-            a_3,
-            a_4
-            | ModulateBy [
-                Length 4, 
-                O[
-                    (3/2, 0, 1, 0),
-                    (1/1, 7, 1, 0),
-                    (1/1, 0, 1, 0),
-                ],
-                O[
-                    (3/2, 0, 1, 0),
-                    (1/1, 8, 1, 0),
-                    (1/1, 0, 1, 0),
-                ],
-                O[
-                    (1/1, 8, 1, 0),
-                    (1/1, 0, 1, 0),
-                ],
-                O[
-                    (1/1, 11, 1, 0),
-                    (1/1, 0, 1, 0),
-                ]
-                | ModulateBy [
-                    std.melt 
-                    | Seq [
-                        Length 10, 
-                        Ta -20, Ta -40, Ta -60 
-                    ]
-                ],
-            ]
-        ], 
-        Seq [
-            b_1 | Length 5/4, 
-            b_2, 
-            b_3,
-            b_4
-            | ModulateBy [
-                Length 5, 
-                O[
-                    (2/1, 0, 1, 0),
-                    (1/1, 4, 1, 0),
-                    (1/1, 0, 1, 0),
-                    (1/2, 0, 1, 0),
-                ] | Length 2,
-                O[
-                    (2/1, 0, 1, 0),
-                    (1/1, 3, 1, 0),
-                    (1/1, 0, 1, 0),
-                    (1/2, 0, 1, 0),
-                ],
-                O[
-                    (2/1, 0, 1, 0),
-                    (1/1, 2, 1, 0),
-                    (1/1, 0, 1, 0),
-                    (1/2, 0, 1, 0),
-                ],
-                O[
-                    (2/1, 0, 1, 0),
-                    (1/1, 1, 1, 0),
-                    (1/1, 0, 1, 0),
-                    (1/2, 0, 1, 0),
-                ]
-                | ModulateBy [
-                    std.melt 
-                    | Seq [
-                        Length 10, 
-                        Ta -20, Ta -40, Ta -60, Ta -80, 
-                        Ta -100
-                    ]
-                ],
-            ]
-            
-        ], 
-        Seq [
-            Tm 0 > FitLength Seq [a_1 | Length 5/4],
-            Tm 0 > FitLength Seq [a_2],
-            c_3,
-            c_4
-        ],
-    ]
-}
-
